+++ lib/routes/app_routes.dart
@@ -1,5 +1,0 @@
-class AppRoutes {
-  static const splash = "/";
-  static const login = "/login";
-  static const dashboard = "/dashboard";
-}